+++ api/models/index.js
@@ -1,53 +1,31 @@
-const { Model, DataTypes } = require("sequelize");
-const sequelize = require("../config/connection");
-const bcrypt = require("bcrypt");
+const User = require("./User");
+const Game = require("./Game");
+const Player = require("./Player");
+const Round = require("./Round");
 
-class User extends Model {
-  checkPassword(loginPw) {
-    return bcrypt.compareSync(loginPw, this.password);
-  }
-}
+User.hasOne(Game, {
+  foreignKey: "user_id",
+});
+Game.hasMany(Player, {
+  foreignKey: "game_id"
+});
+User.hasOne(Player, {
+  foreignKey: "user_id"
+});
+Game.hasMany(Round, {
+  foreignKey: "game_id"
+});
+Round.belongsTo(Game, {
+  foreignKey: "game_id"
+});
+Game.belongsTo(User, {
+  foreignKey: "user_id"
+});
+Player.belongsTo(Game, {
+  foreignKey: "game_id"
+});
+Player.belongsTo(User, {
+  foreignKey: "user_id"
+});
 
-User.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    username: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-      validate: {
-        isEmail: true,
-      },
-    },
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      validate: {
-        len: [8],
-      },
-    },
-  },
-  {
-    hooks: {
-      beforeCreate: async (newUserData) => {
-        newUserData.password = await bcrypt.hash(newUserData.password, 10);
-        return newUserData;
-      },
-    },
-    sequelize,
-    freezeTableName: true,
-    underscored: true,
-    modelName: "user",
-  }
-);
-
-module.exports = User;
+module.exports = {User, Game, Player, Round};
